Relocate code to subtree method
